refactor: split recording_video, DRY daemon spawn, standardize file headers
- Split tools_recording_video.go (788→400+394 LOC): extract file I/O,
  upload parsing, reveal, and listing to tools_recording_video_files.go
- Extract markFailed() and buildDaemonCmd() helpers in bridge.go to
  eliminate duplicated daemon spawn logic (437→408 LOC)
- Standardize 37 file headers from legacy `// Purpose:` format to
  required `// filename.go — Purpose summary.` convention
- Update bridge_detach_contract_test.go for buildDaemonCmd extraction
