+++ cmd/dev-console/cli_output.go
@@ -1,7 +1,5 @@
-// Purpose: Owns cli_output.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // cli_output.go — Output formatters for CLI mode: human, json, csv.
+// Docs: docs/features/feature/observe/index.md
 package main
 
 import (
