+++ cmd/dev-console/connect_mode.go
@@ -1,7 +1,5 @@
-// Purpose: Owns connect_mode.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // connect_mode.go — Connect mode for multi-client MCP support.
+// Docs: docs/features/feature/observe/index.md
 // Enables multiple Claude Code sessions to share a single server.
 package main
