+++ cmd/dev-console/health.go
@@ -1,7 +1,5 @@
-// Purpose: Owns health.go runtime behavior and integration logic.
+// health.go — Health and SLA Metrics (Tier 3.4).
 // Docs: docs/features/feature/observe/index.md
-
-// health.go — Health & SLA Metrics (Tier 3.4).
 // Exposes server uptime, buffer utilization, memory usage, request counts,
 // and error rates via the get_health MCP tool.
 // Design: Thread-safe counters using sync.RWMutex. Memory stats from runtime.
