+++ cmd/dev-console/main_connection.go
@@ -1,7 +1,5 @@
-// Purpose: Owns main_connection.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // main_connection.go — MCP client connection lifecycle: spawn, retry, and zombie recovery.
+// Docs: docs/features/feature/observe/index.md
 package main
 
 import (
