+++ cmd/dev-console/main_connection_diag.go
@@ -1,7 +1,5 @@
-// Purpose: Owns main_connection_diag.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // main_connection_diag.go — Connection diagnostics: health checks, port scanning, and failure analysis.
+// Docs: docs/features/feature/observe/index.md
 package main
 
 import (
