+++ cmd/dev-console/main_connection_mcp.go
@@ -1,7 +1,5 @@
-// Purpose: Owns main_connection_mcp.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // main_connection_mcp.go — MCP daemon mode: HTTP server startup, PID management, and signal handling.
+// Docs: docs/features/feature/observe/index.md
 package main
 
 import (
