+++ cmd/dev-console/platform_errors.go
@@ -1,7 +1,5 @@
-// Purpose: Owns platform_errors.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // platform_errors.go — Platform-aware error messages and process utilities.
+// Docs: docs/features/feature/observe/index.md
 package main
 
 import (
