+++ cmd/dev-console/recording_handlers.go
@@ -1,7 +1,5 @@
-// Purpose: Owns recording_handlers.go runtime behavior and integration logic.
+// recording_handlers.go — MCP tool handlers for Flow Recording and Playback feature.
 // Docs: docs/features/feature/observe/index.md
-
-// recording_handlers.go — MCP tool handlers for Flow Recording & Playback feature.
 // Handles recording_start, recording_stop, recording_get, and playback actions.
 package main
