+++ cmd/dev-console/telemetry_passive.go
@@ -1,6 +1,5 @@
-// Purpose: Owns telemetry_passive.go runtime behavior and integration logic.
+// telemetry_passive.go — Passive telemetry collection and integration logic.
 // Docs: docs/features/feature/observe/index.md
-
 package main
 
 import (
