+++ cmd/dev-console/tools_configure.go
@@ -1,7 +1,5 @@
-// Purpose: Owns tools_configure.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // tools_configure.go — MCP configure tool dispatcher and handlers.
+// Docs: docs/features/feature/observe/index.md
 // Handles session settings: store, load, noise_rule, clear, streaming, recordings, etc.
 //
 // JSON CONVENTION: All fields MUST use snake_case. See .claude/refs/api-naming-standards.md
