+++ cmd/dev-console/tools_generate.go
@@ -1,7 +1,5 @@
-// Purpose: Implements generate tool formats and output assembly.
-// Docs: docs/features/feature/test-generation/index.md
-
 // tools_generate.go — MCP generate tool dispatcher and handlers.
+// Docs: docs/features/feature/test-generation/index.md
 // Handles all generate formats: reproduction, test, pr_summary, sarif, har, csp, sri.
 package main
