+++ cmd/dev-console/tools_interact.go
@@ -1,7 +1,5 @@
-// Purpose: Implements interact tool handlers and browser action orchestration.
-// Docs: docs/features/feature/interact-explore/index.md
-
 // tools_interact.go — MCP interact tool dispatcher and handlers.
+// Docs: docs/features/feature/interact-explore/index.md
 // Handles all browser interaction actions: navigate, execute_js, highlight, state management, etc.
 package main
