+++ cmd/dev-console/tools_session_audit.go
@@ -1,4 +1,4 @@
-// Purpose: Session snapshot adapter and audit-trail recording helpers for MCP tools.
+// tools_session_audit.go — Session snapshot adapter and audit-trail recording helpers for MCP tools.
 package main
 
 import (
