+++ cmd/dev-console/version_check.go
@@ -1,7 +1,5 @@
-// Purpose: Owns version_check.go runtime behavior and integration logic.
-// Docs: docs/features/feature/observe/index.md
-
 // version_check.go — GitHub version checking for update notifications.
+// Docs: docs/features/feature/observe/index.md
 // Checks GitHub for new releases and notifies users of available updates.
 package main
